move system_sync decorators to Transport layer
